Extract `epoch_upgrader` class
